+++ stobu/projectadder.py
@@ -8,10 +8,10 @@ import shutil
 
 
 # My Modules
-from stobu import projectfilechecker as checker
 from stobu.projecteditor import edit_the_chapter, edit_the_episode, edit_the_scene, edit_the_note
 from stobu.projecteditor import edit_the_person, edit_the_stage, edit_the_item, edit_the_word
 from stobu.templatecreator import TemplateCreator
+from stobu.tools import filechecker as checker
 from stobu.tools import pathmanager as ppath
 from stobu.util import assertion
 from stobu.util.fileio import write_file
