+++ stobu/tools/filechecker.py
